# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Todo API
  version: 1.0.0
  description: API for managing smart todos with OIDC authentication
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.smarttodo.com
    description: Production server
paths:
...
  /api/v1/auth/me:
    get:
      summary: Get current user info
      description: Returns information about the authenticated user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/todos:
    get:
      summary: List todos
      description: Get all todos for the authenticated user, optionally filtered by time_horizon and status
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: time_horizon
          in: query
          description: Filter by time horizon
          schema:
            type: string
            enum: [next, soon, later]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, processing, completed]
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodosResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new todo
      description: Create a new todo item
      tags:
        - Todos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/ai/context:
    get:
      summary: Get AI context
      description: Returns the current user's AI context (context summary and preferences)
      tags:
        - AI
      security:
        - bearerAuth: []
      responses:
        '200':
          description: AI context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIContextResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update AI context
      description: Updates the current user's AI context (context summary and/or preferences)
      tags:
        - AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAIContextRequest'
      responses:
        '200':
          description: AI context updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIContextResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OIDC provider
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        provider_id:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
        timestamp:
          type: string
          format: date-time
    CreateTodoRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
...
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        text:
          type: string
        time_horizon:
          type: string
          enum: [now, soon, later]
        status:
          type: string
          enum: [pending, processing, completed]
        metadata:
          $ref: '#/components/schemas/Metadata'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    Metadata:
      type: object
      properties:
        category_tags:
          type: array
          items:
            type: string
        priority:
          type: string
        context:
          type: array
          items:
            type: string
        duration:
          type: string
    TodoResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Todo'
        timestamp:
          type: string
          format: date-time
    TodosResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        timestamp:
          type: string
          format: date-time
...
    AIContextResponse:
      type: object
      properties:
        context_summary:
          type: string
          description: User-driven context summary used for todo classification
        preferences:
          type: object
          additionalProperties: true
          description: User preferences stored as key-value pairs
    UpdateAIContextRequest:
      type: object
      properties:
        context_summary:
          type: string
          description: User-driven context summary to update
        preferences:
          type: object
          additionalProperties: true
          description: User preferences to update (merged with existing)
    Error:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'